Interpreter: shift bits as 64-bit on every platform
OP_BITSHIFTL and OP_BITSHIFTR loaded the left operand into an unsigned
long, but popLong() returns qint64. On Windows, where unsigned long is
32 bits, the value was truncated before the shift and the sizeof(a)*8
guard tripped at 32 places instead of 64, so the same program gave
different answers on Windows than on Linux and macOS.

Use quint64 so the operand keeps all 64 bits and the guard trips at 64
everywhere, matching OP_BINARYAND/OR/NOT which already work in qint64.
Note that on Windows a left shift can now move a bit into the sign
position and return a negative number.

+++ src/core/Interpreter.cpp
@@ -6103,7 +6103,7 @@ fprintf(stderr,"in foreach map %d\n", d->map->data.size());
 				// The standard says: "If the value of the right operand is negative or is greater than
 				// or equal to the width of the promoted left operand, the behavior is undefined."
 					int n = stack->popInt();
-					unsigned long a = stack->popLong();
+					quint64 a = (quint64)stack->popLong();
 					if(n>=0){
 						if(n>=((int)sizeof(a))*8){
 							stack->pushInt(0);
@@ -6128,7 +6128,7 @@ fprintf(stderr,"in foreach map %d\n", d->map->data.size());
 				// The standard says: "If the value of the right operand is negative or is greater than
 				// or equal to the width of the promoted left operand, the behavior is undefined."
 					int n = stack->popInt();
-					unsigned long a = stack->popLong();
+					quint64 a = (quint64)stack->popLong();
 					if(n>=0){
 						if(n>=((int)sizeof(a))*8){
 							stack->pushInt(0);
